Fix Dashboard import paths
