+++ package.json
@@ -15,7 +15,9 @@
     "react-dom": "^18.3.1",
     "jspdf": "^2.5.1",
     "html2canvas": "^1.4.1",
-    "uuid": "^9.0.1"
+    "uuid": "^9.0.1",
+    "react-router-dom": "^6.23.0",
+    "@supabase/supabase-js": "^2.39.7"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.1",
